+++ tests/test_fbs_utils.py
@@ -19,5 +19,31 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from .block_utils import *
-from .feature_scheduler_sim import *
+import math
+
+from lsst.ts.scheduler.driver.driver_target import DriverTarget
+from lsst.ts.scheduler.utils.fbs_utils import make_fbs_observation_from_target
+from lsst.ts.scheduler.utils.test.block_utils import get_test_obs_block
+
+
+def test_make_fbs_observation_from_target() -> None:
+    obs_block = get_test_obs_block()
+    ra = 10.0
+    dec = 20.0
+    target = DriverTarget(
+        observing_block=obs_block,
+        ra_rad=math.radians(ra),
+        dec_rad=math.radians(dec),
+        num_exp=2,
+        exp_times=[15, 15],
+        note="Test",
+    )
+
+    fbs_observation = make_fbs_observation_from_target(target=target)
+
+    print(f"{fbs_observation!r}")
+
+    assert fbs_observation["RA"][0] == target.ra_rad
+    assert fbs_observation["dec"][0] == target.dec_rad
+    assert fbs_observation["note"][0] == target.note
+    assert fbs_observation["exptime"][0] == sum(target.exp_times)